size_t for sharedSize

+++ include/linear.h
@@ -29,7 +29,7 @@ public:
   //  int zderiv(MomentsG *G);
 
   dim3 dimGrid, dimBlock, dG, dB, dGs, dBs, dimGridh, dimBlockh, dB_all, dG_all;
-  int sharedSize;
+  size_t sharedSize;
   
  private:
   bool ks;
@@ -65,7 +65,7 @@ public:
   void get_max_frequency(double* wmax);
 
   dim3 dimGrid, dimBlock, dG, dB, dGs, dBs, dimGridh, dimBlockh, dB_all, dG_all;
-  int sharedSize;
+  size_t sharedSize;
   
  private:
